% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gelman_good_proportion}
\alias{gelman_good_proportion}
\title{Calculate the proportion of variables in an mcmc object that have "good" gelman.diag}
\usage{
gelman_good_proportion(x, psrf_limit = 1.1)
}
\arguments{
\item{x}{an mcmc object}

\item{psrf_limit}{numeric the limit to classify the diag as good}
}
\value{
A numeric vector of length two. The first number is the proportion of
gelman.diag() <= `psrf_limit` the second number is the number of variables
in `x`.
}
\description{
Calculate the proportion of variables in an mcmc object that have an
[coda::gelman.diag()] value below or equal psrf_limit (defalt = 1.10)
}
\examples{
require(coda)
data(line) # data from package coda
gelman_good_proportion(line)

}
