% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_blackbox_coda}
\alias{read_blackbox_coda}
\title{Read coda saved in BlackBox txt format and return a mcmc.list}
\usage{
read_blackbox_coda(
  directory = ".",
  index = "index.txt",
  chain1 = "chain1.txt",
  chain2 = "chain2.txt"
)
}
\arguments{
\item{directory}{character default current directory "."}

\item{index}{character name of index file default "index.txt"}

\item{chain1}{character name of file with first chain default "chain1.txt"}

\item{chain2}{character name of file with second chain default "chain2.txt"}
}
\value{
return an object of class mcmc.list
}
\description{
The function will read three files saved as text from BlackBox with Inference/Samples.
All three files must be in the same directory.
}
