% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_CU}
\alias{read_CU}
\title{Read the CU (smolt estimate coda) from the smolt model results}
\usage{
read_CU(river, year, species)
}
\arguments{
\item{species}{}
}
\value{
Returns a data frame with the columns iteration, chain1 and chain2.
}
\description{
Read the CU (smolt estimate coda) from the smolt model results
}
