% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_model_result}
\alias{read_model_result}
\title{Read summarized results from smolt model}
\usage{
read_model_result(directory = ".", resultfile = "model_results.xlsx")
}
\arguments{
\item{directory}{character default current directory "."}

\item{resultfile}{character Default "model_results.xlsx" in current directory.}
}
\value{
Returns a data frame with the results from a smolt trap in river in year for species.
}
\description{
Read summarized results from smolt model
}
