% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_rdata}
\alias{read_rdata}
\title{Read data used for Smolt mark recaputure model}
\usage{
read_rdata(directory = ".", rdatafile = "RData_dump.RData")
}
\arguments{
\item{directory}{character default current directory "."}

\item{rdatafile}{character Default "RData_dump.RData" in current directory.}
}
\value{
Returns a data structure with the data from a smolt trap in river in year for species.
}
\description{
Read the rdata file saved by söte2bugs. If vectors c, m or r contains -9
they are replaced with NA.
}
