% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_from_mcmc}
\alias{select_from_mcmc}
\title{Select a subset of variables from a coda object}
\usage{
select_from_mcmc(x, prefixes)
}
\arguments{
\item{x}{a `mcmc` object}

\item{prefixes}{character vector}
}
\value{
A `mcmc` object with only the variables
}
\description{
Select a subset of variables from an [coda::mcmc()] or [coda::mcmc.list()]
object. Variables with names starting with one of the strings in `prefixes`
are selected.
}
\examples{
require(coda)
data(line) # data from package coda
coda::varnames(line)
new_line <- select_from_mcmc(line, c("al", "si"))
coda::varnames(new_line)

}
