% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startsWith_multiple}
\alias{startsWith_multiple}
\title{Selects strings that match one on several prefixes}
\usage{
startsWith_multiple(strings, prefixes)
}
\arguments{
\item{strings}{character vector to match}

\item{prefixes}{character vector with prefixes}
}
\value{
A logical vector with the same length as `strings`
}
\description{
Match strings for the prefixes using base::startsWith(). Return a logical
vector that indicate all matches.
}
\examples{
x1 <- c("Foobar", "bla bla", "something", "another", "blu", "brown", "blau blüht der Enzian")
p <- c("Foo", "bl")
startsWith_multiple(x1, p)

}
