% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_CU}
\alias{read_CU}
\title{Read the CU (smolt estimate coda) from the smolt model results.}
\usage{
read_CU(directory = ".", resultfile = "model_results.xlsx")
}
\arguments{
\item{directory}{character default current directory "."}

\item{resultfile}{character filename of saved results. Default "model_results.xlsx"}
}
\value{
Returns a data frame with the columns iteration, chain1 and chain2. Chain1 and chain2 are
samples (posterior) of the estimated CU (total smolt run).
}
\description{
Read the CU (smolt estimate coda) from the smolt model results.
}
