% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_model_result}
\alias{read_model_result}
\title{Read results from smolt model}
\usage{
read_model_result(river, year, species)
}
\arguments{
\item{species}{}
}
\value{
Returns a data frame with the results from a smolt trap in river in year for species.
}
\description{
Read results from smolt model
}
