% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloaders.R
\name{read_rdata}
\alias{read_rdata}
\title{Read data used for Smolt mark recaputure model}
\usage{
read_rdata(river, year, species)
}
\arguments{
\item{species}{}
}
\value{
Returns a data structure with the data from a smolt trap in river in year for species.
}
\description{
Read data used for Smolt mark recaputure model
}
